Змінив СMake Тепер файл для пайтону робочий , пофіксив і змінив структуру деяких методів та додав нові , для правильної роботи з пайтоном

+++ Numerical_battles_cpp/Enemy.cpp
@@ -46,7 +46,7 @@ double Enemy::getNumber() {
     return this->number;
 }
 double Enemy::setNumber(int number) {
-    this->number += number;
+    this->number += (double)number;
     return this->number;
 }
